refactor: refactor content management page
