+++ components/SingleContentForm.jsx
@@ -4,6 +4,9 @@ import { useState } from "react";
 import { Upload, Tag } from "lucide-react";
 import isValidDate from "@/components/dateValidation"
 import useToastStore from '@/lib/store/toast';
+import {
+  getStatusColor
+} from "@/utils/contentHelpers";
 export default function SingleContentForm({
   setIsModalOpen,
   currentContent,
@@ -22,14 +25,14 @@ export default function SingleContentForm({
   const [isWeeklyContentChecked, setIsWeeklyContentChecked] = useState(
     !!currentContent?.isWeeklyContent
   );
-    const [studentDate, setStudentDate] = useState(currentContent?.publishDateStudent
-      ? new Date(currentContent.publishDateStudent).toISOString().split("T")[0]
-      : "");
-    
-    const [teacherDate, setTeacherDate] = useState(currentContent?.publishDateTeacher
-      ? new Date(currentContent.publishDateTeacher).toISOString().split("T")[0]
-      : "");
-      const { showToast } = useToastStore();
+  const [studentDate, setStudentDate] = useState(currentContent?.publishDateStudent
+    ? new Date(currentContent.publishDateStudent).toISOString().split("T")[0]
+    : "");
+
+  const [teacherDate, setTeacherDate] = useState(currentContent?.publishDateTeacher
+    ? new Date(currentContent.publishDateTeacher).toISOString().split("T")[0]
+    : "");
+  const { showToast } = useToastStore();
 
 
   const uploadFileToR2 = async (file) => {
@@ -96,7 +99,7 @@ export default function SingleContentForm({
       }
     }
 
-    if ( !fileUrl) {
+    if (!fileUrl) {
       showToast("Lütfen içerik dosyası ekleyin.", "error");
       setIsUploading(false);
       return;
@@ -291,22 +294,22 @@ export default function SingleContentForm({
                         type="date"
                         name="publishDateStudent"
                         id="publishDateStudent"
-                        min={new Date().toISOString().split("T")[0]} 
+                        min={new Date().toISOString().split("T")[0]}
                         value={studentDate}
                         onChange={(e) => {
                           const selectedDate = e.target.value;
-                      
+
                           if (teacherDate) {
                             const student = new Date(selectedDate);
                             const teacher = new Date(teacherDate);
-                      
+
                             if (student < teacher) {
                               showToast("Öğrenci yayın tarihi, öğretmen tarihinden önce olamaz!", "error");
                               setStudentDate("");
                               return;
                             }
                           }
-                      
+
                           setStudentDate(selectedDate);
                         }}
                         className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3"
@@ -322,9 +325,9 @@ export default function SingleContentForm({
                         type="date"
                         name="publishDateTeacher"
                         id="publishDateTeacher"
-                        min={new Date().toISOString().split("T")[0]} 
+                        min={new Date().toISOString().split("T")[0]}
                         value={teacherDate}
-                     
+
                         onChange={(e) => {
                           const selectedDate = e.target.value;
                           if (studentDate) {
@@ -336,10 +339,10 @@ export default function SingleContentForm({
                               return;
                             }
                           }
-                        
+
                           setTeacherDate(selectedDate);
                         }}
-                        
+
                         className="mt-1 block w-full border border-gray-300 rounded-md shadow-sm py-2 px-3"
                       />
                     </div>
@@ -384,7 +387,7 @@ export default function SingleContentForm({
                           type="date"
                           name="weeklyContentStartDate"
                           id="weeklyContentStartDate"
-                          min={new Date().toISOString().split("T")[0]} 
+                          min={new Date().toISOString().split("T")[0]}
                           defaultValue={
                             isValidDate(currentContent?.weeklyContentStartDate)
                               ? new Date(currentContent.weeklyContentStartDate)
